requirements.txt
alembic==1.16.4
idna==3.3
annotated-types==0.7.0
asgiref==3.5.0
attrs==25.3.0
bcrypt==3.2.0
blinker==1.4
certifi==2025.4.26
chardet==4.0.0
exceptiongroup==1.2.2
fastapi
fasteners==0.14.1
httplib2==0.20.2
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
oauthlib==3.2.0
odfpy==1.4.1
olefile==0.46
outcome==1.3.0.post0
pandas==2.2.3
paramiko==2.9.3
passlib==1.7.4
pexpect==4.8.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pyasn1==0.6.1
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydantic[email]
pydub==0.25.1
pygame==2.6.1
PyJWT==2.3.0
pytest==8.4.1
pytest-html==4.1.1
pytest-metadata==3.1.1
httpx
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
PySocks==1.7.1
python-dateutil==2.9.0.post
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2022.1
pyxdg==0.27
rsa==4.9.1
requests
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.41
starlette==0.47.2
tomli==2.2.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==1.26.5
uvicorn==0.15.0